+++ source/game_sa/COcclusion.cpp
@@ -1,5 +1,7 @@
 #include "StdInc.h"
 
+#include "COcclusion.h"
+
 COccluder(&COcclusion::aInteriorOccluders)[MAX_INTERIOR_OCCLUDERS] = *(COccluder(*)[MAX_INTERIOR_OCCLUDERS])0xC73CC8;
 COccluder(&COcclusion::aOccluders)[MAX_MAP_OCCLUDERS] = *(COccluder(*)[MAX_MAP_OCCLUDERS])0xC73FA0;
 CActiveOccluder(&COcclusion::aActiveOccluders)[MAX_ACTIVE_OCCLUDERS] = *(CActiveOccluder(*)[MAX_ACTIVE_OCCLUDERS])0xC78610;
@@ -29,13 +31,6 @@ void COcclusion::InjectHooks()
     ReversibleHooks::Install("COcclusion", "IsPositionOccluded", 0x7200B0, &COcclusion::IsPositionOccluded);
     ReversibleHooks::Install("COcclusion", "OccluderHidesBehind", 0x71E080 , &COcclusion::OccluderHidesBehind);
     ReversibleHooks::Install("COcclusion", "ProcessBeforeRendering", 0x7201C0, &COcclusion::ProcessBeforeRendering);
-
-    ReversibleHooks::Install("COccluder", "ProcessOneOccluder", 0x71E5D0, &COccluder::ProcessOneOccluder);
-    ReversibleHooks::Install("COccluder", "ProcessLineSegment", 0x71E130, &COccluder::ProcessLineSegment);
-    ReversibleHooks::Install("COccluder", "NearCamera", 0x71F960, &COccluder::NearCamera);
-
-    ReversibleHooks::Install("CActiveOccluder", "IsPointWithinOcclusionArea", 0x71E580, &CActiveOccluder::IsPointWithinOcclusionArea);
-    ReversibleHooks::Install("CActiveOccluder", "IsPointBehindOccluder", 0x71FA40, &CActiveOccluder::IsPointBehindOccluder);
 }
 
 // 0x71DCA0
@@ -278,275 +273,4 @@ void COcclusion::ProcessBeforeRendering()
             }
         }
     }
-}
-
-bool CActiveOccluder::IsPointWithinOcclusionArea(float fX, float fY, float fRadius)
-{
-    if (m_cLinesCount <= 0)
-        return true;
-
-    for (auto i = 0; i < m_cLinesCount; ++i) {
-        auto& line = m_aLines[i];
-        if (!IsPointInsideLine(line.m_vecOrigin.x, line.m_vecOrigin.y, line.m_vecDirection.x, line.m_vecDirection.y, fX, fY, fRadius))
-            return false;
-    }
-
-    return true;
-}
-
-bool CActiveOccluder::IsPointBehindOccluder(CVector vecPos, float fRadius)
-{
-    if (m_cNumVectors <= 0)
-        return true;
-
-    for (auto i = 0; i < m_cNumVectors; ++i) {
-        const auto& vecCamPos = TheCamera.GetPosition();
-
-        auto fPosDotVec = DotProduct(vecPos, m_aVectors[i]) - m_afRadiuses[i];
-        auto fCamDotVec = DotProduct(vecCamPos, m_aVectors[i]) - m_afRadiuses[i];
-
-        if (fCamDotVec * fPosDotVec >= 0.0F || fabs(fPosDotVec) < fRadius)
-            return false;
-    }
-
-    return true;
-}
-
-// 0x71E5D0
-bool COccluder::ProcessOneOccluder(CActiveOccluder* pActiveOccluder)
-{
-    pActiveOccluder->m_cLinesCount = 0;
-    auto vecPos = CVector(m_wMidX / 4.0F, m_wMidY / 4.0F, m_wMidZ / 4.0F);
-    float temp1, temp2;
-
-    if (!CalcScreenCoors(vecPos, &COcclusion::gCenterOnScreen, &temp1, &temp2) || COcclusion::gCenterOnScreen.z < -150.0F || COcclusion::gCenterOnScreen.z > 300.0F)
-        return false;
-
-    auto fMagnitude = CVector(m_wWidth / 4.0F, m_wLength / 4.0F, m_wHeight / 4.0F).Magnitude();
-    pActiveOccluder->m_wDepth = COcclusion::gCenterOnScreen.z - fMagnitude;
-
-    auto matRotX = CMatrix();
-    auto matRotY = CMatrix();
-    auto matRotZ = CMatrix();;
-    auto matTransform = CMatrix();
-
-    matRotX.SetRotateX(m_cRotX * PI / 128.0F);
-    matRotY.SetRotateY(m_cRotY * PI / 128.0F);
-    matRotZ.SetRotateZ(m_cRotZ * PI / 128.0F);
-    matTransform = (matRotY * matRotX) * matRotZ;
-
-    COcclusion::gMinXInOccluder = 999999.88F;
-    COcclusion::gMinYInOccluder = 999999.88F;
-    COcclusion::gMaxXInOccluder = -999999.88F;
-    COcclusion::gMaxYInOccluder = -999999.88F;
-
-    if (   m_wLength * 0.25F != 0.0F
-        && m_wWidth  * 0.25F != 0.0F
-        && m_wHeight * 0.25F != 0.0F) {
-        auto vecWidth = CVector(m_wWidth / 8.0F, 0.0F, 0.0F);
-        auto vecTransWidth = matTransform * vecWidth;
-
-        auto vecLength = CVector(0.0F, m_wLength / 8.0F, 0.0F);
-        auto vecTransLength = matTransform * vecLength;
-
-        auto vecHeight = CVector(0.0F, 0.0F, m_wHeight / 8.0F);
-        auto vecTransHeight = matTransform * vecHeight;
-
-        CVector aVecArr[6]{
-            vecTransLength, -vecTransLength,
-            vecTransWidth, -vecTransWidth,
-            vecTransHeight, -vecTransHeight
-        };
-
-        // Figure out if we see the front or back of a face
-        const auto& vecCamPos = TheCamera.GetPosition();
-        bool abOnScreen[6]{
-            DotProduct((vecPos + aVecArr[0] - vecCamPos), aVecArr[0]) < 0.0F,
-            DotProduct((vecPos + aVecArr[1] - vecCamPos), aVecArr[1]) < 0.0F,
-            DotProduct((vecPos + aVecArr[2] - vecCamPos), aVecArr[2]) < 0.0F,
-            DotProduct((vecPos + aVecArr[3] - vecCamPos), aVecArr[3]) < 0.0F,
-            DotProduct((vecPos + aVecArr[4] - vecCamPos), aVecArr[4]) < 0.0F,
-            DotProduct((vecPos + aVecArr[5] - vecCamPos), aVecArr[5]) < 0.0F
-        };
-
-        // Calculating vertices of a box
-        COcclusion::gOccluderCoors[0] = vecPos + aVecArr[0] + aVecArr[2] + aVecArr[4];
-        COcclusion::gOccluderCoors[1] = vecPos + aVecArr[1] + aVecArr[2] + aVecArr[4];
-        COcclusion::gOccluderCoors[2] = vecPos + aVecArr[0] + aVecArr[3] + aVecArr[4];
-        COcclusion::gOccluderCoors[3] = vecPos + aVecArr[1] + aVecArr[3] + aVecArr[4];
-        COcclusion::gOccluderCoors[4] = vecPos + aVecArr[0] + aVecArr[2] + aVecArr[5];
-        COcclusion::gOccluderCoors[5] = vecPos + aVecArr[1] + aVecArr[2] + aVecArr[5];
-        COcclusion::gOccluderCoors[6] = vecPos + aVecArr[0] + aVecArr[3] + aVecArr[5];
-        COcclusion::gOccluderCoors[7] = vecPos + aVecArr[1] + aVecArr[3] + aVecArr[5];
-
-        for (auto i = 0; i < 8; ++i) {
-            COcclusion::gOccluderCoorsValid[i] = CalcScreenCoors(COcclusion::gOccluderCoors[i], &COcclusion::gOccluderCoorsOnScreen[i], &temp1, &temp2);
-        }
-
-        // Between two differently facing sides we see an edge, so process those
-        if (   (abOnScreen[0] == abOnScreen[2] || !ProcessLineSegment(0, 4, pActiveOccluder))
-            && (abOnScreen[0] == abOnScreen[3] || !ProcessLineSegment(2, 6, pActiveOccluder))
-            && (abOnScreen[0] == abOnScreen[4] || !ProcessLineSegment(0, 2, pActiveOccluder))
-            && (abOnScreen[0] == abOnScreen[5] || !ProcessLineSegment(4, 6, pActiveOccluder))
-            && (abOnScreen[1] == abOnScreen[2] || !ProcessLineSegment(1, 5, pActiveOccluder))
-            && (abOnScreen[1] == abOnScreen[3] || !ProcessLineSegment(3, 7, pActiveOccluder))
-            && (abOnScreen[1] == abOnScreen[4] || !ProcessLineSegment(1, 3, pActiveOccluder))
-            && (abOnScreen[1] == abOnScreen[5] || !ProcessLineSegment(5, 7, pActiveOccluder))
-            && (abOnScreen[2] == abOnScreen[4] || !ProcessLineSegment(0, 1, pActiveOccluder))
-            && (abOnScreen[3] == abOnScreen[4] || !ProcessLineSegment(2, 3, pActiveOccluder))
-            && (abOnScreen[3] == abOnScreen[5] || !ProcessLineSegment(6, 7, pActiveOccluder))
-            && (abOnScreen[2] == abOnScreen[5] || !ProcessLineSegment(4, 5, pActiveOccluder))
-            && SCREEN_WIDTH * 0.15F <= COcclusion::gMaxXInOccluder - COcclusion::gMinXInOccluder
-            && SCREEN_HEIGHT * 0.1F <= COcclusion::gMaxYInOccluder - COcclusion::gMinYInOccluder) {
-
-            pActiveOccluder->m_cNumVectors = 0;
-            for (auto i = 0; i < 6; ++i) {
-                if (abOnScreen[i]) {
-                    auto vecNormalised = CVector(aVecArr[i]);
-                    vecNormalised.Normalise();
-                    auto vecScreenPos = vecPos + aVecArr[i];
-
-                    pActiveOccluder->m_aVectors[pActiveOccluder->m_cNumVectors] = vecNormalised;
-                    pActiveOccluder->m_afRadiuses[pActiveOccluder->m_cNumVectors] = DotProduct(vecScreenPos, vecNormalised);
-                    ++pActiveOccluder->m_cNumVectors;
-                }
-            }
-            return true;
-        }
-
-        return false;
-    }
-
-    CVector vec1, vec2;
-    if (m_wLength * 0.25F == 0.0F) {
-        auto vecWidth = CVector(m_wWidth / 8.0F, 0.0F, 0.0F);
-        vec1 = matTransform * vecWidth;
-
-        auto vecHeight = CVector(0.0F, 0.0F, m_wHeight / 8.0F);
-        vec2 = matTransform * vecHeight;
-    }
-    else if (m_wWidth * 0.25F == 0.0F) {
-        auto vecLength = CVector(0.0F, m_wLength / 8.0F, 0.0F);
-        vec1 = matTransform * vecLength;
-
-        auto vecHeight = CVector(0.0F, 0.0F, m_wHeight / 8.0F);
-        vec2 = matTransform * vecHeight;
-    }
-    else if (m_wHeight * 0.25F == 0.0F) {
-        auto vecLength = CVector(0.0F, m_wLength / 8.0F, 0.0F);
-        vec1 = matTransform * vecLength;
-
-        auto vecWidth = CVector(m_wWidth / 8.0F, 0.0F, 0.0F);
-        vec2 = matTransform * vecWidth;
-    }
-
-    COcclusion::gOccluderCoors[0] = vecPos + vec1 + vec2;
-    COcclusion::gOccluderCoors[1] = vecPos - vec1 + vec2;
-    COcclusion::gOccluderCoors[2] = vecPos - vec1 - vec2;
-    COcclusion::gOccluderCoors[3] = vecPos + vec1 - vec2;
-
-    for (auto i = 0; i < 4; ++i) {
-        COcclusion::gOccluderCoorsValid[i] = CalcScreenCoors(COcclusion::gOccluderCoors[i], &COcclusion::gOccluderCoorsOnScreen[i], &temp1, &temp2);
-    }
-
-    if (   !ProcessLineSegment(0, 1, pActiveOccluder)
-        && !ProcessLineSegment(1, 2, pActiveOccluder)
-        && !ProcessLineSegment(2, 3, pActiveOccluder)
-        && !ProcessLineSegment(3, 0, pActiveOccluder)
-        && SCREEN_WIDTH  * 0.1F  <= COcclusion::gMaxXInOccluder - COcclusion::gMinXInOccluder
-        && SCREEN_HEIGHT * 0.07F <= COcclusion::gMaxYInOccluder - COcclusion::gMinYInOccluder){
-
-        auto vecCross = CrossProduct(vec1, vec2);
-        vecCross.Normalise();
-
-        pActiveOccluder->m_aVectors[0] = vecCross;
-        pActiveOccluder->m_afRadiuses[0] = DotProduct(vecCross, vecPos);
-        pActiveOccluder->m_cNumVectors = 1;
-
-        return true;
-    }
-
-    return false;
-}
-
-// 0x71E130
-bool COccluder::ProcessLineSegment(int32 iIndFrom, int32 iIndTo, CActiveOccluder* pActiveOccluder)
-{
-    if (!COcclusion::gOccluderCoorsValid[iIndFrom] && !COcclusion::gOccluderCoorsValid[iIndTo])
-        return true;
-
-    CVector vecScreenFrom, vecScreenTo;
-    if (COcclusion::gOccluderCoorsValid[iIndFrom]) {
-        vecScreenFrom = COcclusion::gOccluderCoorsOnScreen[iIndFrom];
-    }
-    else {
-        auto fFromDepth = fabs((TheCamera.m_mViewMatrix * COcclusion::gOccluderCoors[iIndFrom]).z - 1.1F);
-        auto fToDepth = fabs((TheCamera.m_mViewMatrix * COcclusion::gOccluderCoors[iIndTo]).z - 1.1F);
-
-        auto fProgress = fToDepth / (fFromDepth + fToDepth);
-        vecScreenFrom = (1.0F - fProgress) * COcclusion::gOccluderCoors[iIndTo];
-        vecScreenTo = vecScreenFrom + fProgress * COcclusion::gOccluderCoors[iIndFrom];
-
-        float fTemp1, fTemp2;
-        if (!CalcScreenCoors(vecScreenTo, &vecScreenFrom, &fTemp1, &fTemp2))
-            return true;
-    }
-
-    if (COcclusion::gOccluderCoorsValid[iIndTo]) {
-        vecScreenTo = COcclusion::gOccluderCoorsOnScreen[iIndTo];
-    }
-    else {
-        auto fFromDepth = fabs((TheCamera.m_mViewMatrix * COcclusion::gOccluderCoors[iIndFrom]).z - 1.1F);
-        auto fToDepth = fabs((TheCamera.m_mViewMatrix * COcclusion::gOccluderCoors[iIndTo]).z - 1.1F);
-
-        auto fProgress = fToDepth / (fFromDepth + fToDepth);
-        auto vecFrom = (1.0F - fProgress) * COcclusion::gOccluderCoors[iIndTo];
-        auto vecTo = vecFrom + fProgress * COcclusion::gOccluderCoors[iIndFrom];
-
-        float fTemp1, fTemp2;
-        if (!CalcScreenCoors(vecTo, &vecScreenTo, &fTemp1, &fTemp2))
-            return true;
-    }
-
-    COcclusion::gMinXInOccluder = std::min({ COcclusion::gMinXInOccluder, vecScreenFrom.x, vecScreenTo.x });
-    COcclusion::gMaxXInOccluder = std::max({ COcclusion::gMaxXInOccluder, vecScreenFrom.x, vecScreenTo.x });
-    COcclusion::gMinYInOccluder = std::min({ COcclusion::gMinYInOccluder, vecScreenFrom.y, vecScreenTo.y });
-    COcclusion::gMaxYInOccluder = std::max({ COcclusion::gMaxYInOccluder, vecScreenFrom.y, vecScreenTo.y });
-
-    auto fXSize = vecScreenTo.x - vecScreenFrom.x;
-    auto fYSize = vecScreenTo.y - vecScreenFrom.y;
-    auto fFromX = vecScreenFrom.x;
-    auto fFromY = vecScreenFrom.y;
-
-    if (!IsPointInsideLine(vecScreenFrom.x, vecScreenFrom.y, fXSize, fYSize, COcclusion::gCenterOnScreen.x, COcclusion::gCenterOnScreen.y, 0.0F)) {
-        fFromX = vecScreenFrom.x + fXSize;
-        fFromY = vecScreenFrom.y + fYSize;
-        fXSize = -fXSize;
-        fYSize = -fYSize;
-    }
-
-    auto& pCurLine = pActiveOccluder->m_aLines[pActiveOccluder->m_cLinesCount];
-    pCurLine.m_fLength = CVector2D(fXSize, fYSize).Magnitude();
-
-    auto fRecip = 1.0F / pCurLine.m_fLength;
-    pCurLine.m_vecOrigin.Set(fFromX, fFromY);
-    pCurLine.m_vecDirection.Set(fRecip * fXSize, fRecip * fYSize);
-
-    if (DoesInfiniteLineTouchScreen(fFromX, fFromY, pCurLine.m_vecDirection.x, pCurLine.m_vecDirection.y)) {
-        ++pActiveOccluder->m_cLinesCount;
-        return false;
-    }
-
-    return !IsPointInsideLine(fFromX, fFromY, pCurLine.m_vecDirection.x, pCurLine.m_vecDirection.y, SCREEN_WIDTH * 0.5F, SCREEN_HEIGHT * 0.5F, 0.0F);
-}
-
-// 0x71F960
-bool COccluder::NearCamera()
-{
-    auto fSize = std::max(m_wLength / 4.0F, m_wWidth / 4.0F);
-    const auto& vecCamPos = TheCamera.GetPosition();
-    auto vecPos = CVector(m_wMidX, m_wMidY, m_wMidZ) / 4.0F;
-
-    auto fDist = DistanceBetweenPoints(vecPos, vecCamPos);
-    return (fDist - fSize / 2.0F) < 250.0F;
 }
