+++ source/game_sa/common.cpp
@@ -1116,10 +1116,14 @@ void LittleTest() {
     ++g_nNumIm3dDrawCalls;
 }
 
+// used only in COccluder::ProcessLineSegment
+// 0x71DB80
 bool DoesInfiniteLineTouchScreen(float fX, float fY, float fXDir, float fYDir) {
     return plugin::CallAndReturn<bool, 0x71DB80, float, float, float, float>(fX, fY, fXDir, fYDir);
 }
 
+// Used only in COcclusion, COccluder, CActiveOccluder
+// 0x71E050
 bool IsPointInsideLine(float fLineX, float fLineY, float fXDir, float fYDir, float fPointX, float fPointY, float fTolerance) {
     return (fPointX - fLineX) * fYDir - (fPointY - fLineY) * fXDir >= fTolerance;
 }
